requirements.txt
Flask>=3.0.0
Pillow>=10.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scikit-image>=0.21.0
torch>=2.0.0
torchvision>=0.15.0
umap-learn>=0.5.4
matplotlib>=3.7.0
plotly>=5.15.0
opencv-python>=4.8.0
markdown>=3.5
